+++ saqc/funcs/outliers.py
@@ -39,6 +39,37 @@ if TYPE_CHECKING:
     from saqc import SaQC
 
 
+def _stray(partition, min_periods, iter_start, alpha):
+    sample_size = len(partition)
+    flag_index = pd.Index([])
+    if len(partition) == 0 or sample_size < min_periods:
+        return flag_index
+    if isinstance(partition, np.ndarray):
+        idx = np.arange(len(partition))
+    else:
+        idx = partition.index
+        partition = partition.values
+    sorted_i = partition.argsort()
+    resids = partition[sorted_i]
+    gaps = np.append(0, np.diff(resids))
+
+    tail_size = int(max(min(np.floor(sample_size / 4), 50), 2))
+    tail_indices = np.arange(2, tail_size + 1)
+
+    i_start = int(max(np.floor(sample_size * iter_start), 1))
+    ghat = np.array([np.nan] * sample_size)
+
+    for i in range(i_start, sample_size):
+        ghat[i] = sum((tail_indices / (tail_size - 1)) * gaps[i - tail_indices + 1])
+
+    log_alpha = np.log(1 / alpha)
+    for iter_index in range(i_start, sample_size):
+        if gaps[iter_index] > log_alpha * ghat[iter_index]:
+            flag_index = idx[sorted_i[iter_index:]]
+            break
+    return flag_index
+
+
 class OutliersMixin:
     @staticmethod
     def _validateLOF(algorithm, n, p, density):
@@ -508,32 +539,8 @@ class OutliersMixin:
 
         # calculate flags for every window
         for _, partition in partitions:
-            sample_size = len(partition)
-
-            if partition.empty or sample_size < min_periods:
-                continue
-
-            sorted_i = partition.values.argsort()
-            resids = partition.values[sorted_i]
-            gaps = np.append(0, np.diff(resids))
-
-            tail_size = int(max(min(np.floor(sample_size / 4), 50), 2))
-            tail_indices = np.arange(2, tail_size + 1)
-
-            i_start = int(max(np.floor(sample_size * iter_start), 1) + 1)
-            ghat = np.array([np.nan] * sample_size)
-
-            for i in range(i_start - 1, sample_size):
-                ghat[i] = sum(
-                    (tail_indices / (tail_size - 1)) * gaps[i - tail_indices + 1]
-                )
-
-            log_alpha = np.log(1 / alpha)
-            for iter_index in range(i_start - 1, sample_size):
-                if gaps[iter_index] > log_alpha * ghat[iter_index]:
-                    index = partition.index[sorted_i[iter_index:]]
-                    self._flags[index, field] = flag
-                    break
+            index = _stray(partition, min_periods, iter_start, alpha)
+            self._flags[index, field] = flag
 
         return self